+++ footer.php
@@ -26,7 +26,7 @@
 
 		</div>
 
-		<?php // all js scripts are loaded in library/bones.php ?>
+		<?php // all js scripts are loaded in assets/bones.php ?>
 		<?php wp_footer(); ?>
 
 	</body>