+++ functions.php
@@ -9,10 +9,22 @@ sidebars, comments, etc.
 */
 
 // LOAD BONES CORE (if you remove this, the theme will break)
-require_once( 'library/bones.php' );
+require_once( 'assets/bones.php' );
+
+// LOAD & ENQUEUE SCRIPTS AND STYLES
+require_once( 'assets/scripts.php' );
 
 // CUSTOMIZE THE WORDPRESS ADMIN (off by default)
-// require_once( 'library/admin.php' );
+require_once( 'assets/admin.php' );
+
+// AUTOLOAD THE COMPOSER INCLUDES 
+require_once( 'includes/vendor/autoload.php' );
+
+// INCLUDE THE CUSTOM POST TYPE CLASS
+require_once( 'includes/custom-post-types.php' );
+
+// INCLUDE THE OPTIONS PAGE (ACF)
+require_once( 'includes/options-page.php' );
 
 /*********************
 LAUNCH BONES
@@ -22,13 +34,10 @@ Let's get everything up and running.
 function bones_ahoy() {
 
   //Allow editor style.
-  add_editor_style( get_stylesheet_directory_uri() . '/library/css/editor-style.css' );
+  add_editor_style( get_stylesheet_directory_uri() . '/assets/css/editor-style.css' );
 
   // let's get language support going, if you need it
-  load_theme_textdomain( 'bonestheme', get_template_directory() . '/library/translation' );
-
-  // USE THIS TEMPLATE TO CREATE CUSTOM POST TYPES EASILY
-  require_once( 'library/custom-post-type.php' );
+  load_theme_textdomain( 'bonestheme', get_template_directory() . '/assets/translation' );
 
   // launching operation cleanup
   add_action( 'init', 'bones_head_cleanup' );
@@ -210,7 +219,7 @@ function bones_comments( $comment, $args, $depth ) {
           // create variable
           $bgauthemail = get_comment_author_email();
         ?>
-        <img data-gravatar="http://www.gravatar.com/avatar/<?php echo md5( $bgauthemail ); ?>?s=40" class="load-gravatar avatar avatar-48 photo" height="40" width="40" src="<?php echo get_template_directory_uri(); ?>/library/images/nothing.gif" />
+        <img data-gravatar="http://www.gravatar.com/avatar/<?php echo md5( $bgauthemail ); ?>?s=40" class="load-gravatar avatar avatar-48 photo" height="40" width="40" src="<?php echo get_template_directory_uri(); ?>/assets/images/nothing.gif" />
         <?php // end custom gravatar call ?>
         <?php printf(__( '<cite class="fn">%1$s</cite> %2$s', 'bonestheme' ), get_comment_author_link(), edit_comment_link(__( '(Edit)', 'bonestheme' ),'  ','') ) ?>
         <time datetime="<?php echo comment_time('Y-m-j'); ?>"><a href="<?php echo htmlspecialchars( get_comment_link( $comment->comment_ID ) ) ?>"><?php comment_time(__( 'F jS, Y', 'bonestheme' )); ?> </a></time>
@@ -243,5 +252,10 @@ function bones_fonts() {
 }
 
 add_action('wp_enqueue_scripts', 'bones_fonts');
+
+
+add_filter('show_admin_bar', '__return_false');
+
+register_new_royalslider_files(1);
 
 /* DON'T DELETE THIS CLOSING TAG */ ?>