+++ includes/custom-post-types.php
@@ -1,0 +1,9 @@
+<?php
+
+// // Events Custom Post Type ------
+// $events = new CPT('event', array(
+//   'supports' => array( 'title', 'editor', 'thumbnail' )
+// ));
+// $events->menu_icon('dashicons-calendar-alt');
+// // $events->flush();
+// // Events Custom Post Type ------